+++ src/components/teacherDashboard/addTopic/AddTopic.module.css
@@ -2,7 +2,7 @@
 	display: flex;
 	flex-direction: row;
 	justify-content: space-between;
-	gap: 1rem;
+	gap: 2px;
 }
 
 .add-btn {
