+++ src/components/teacherDashboard/addTopic/index.ts
@@ -1,0 +1,1 @@
+export { AddTopic } from "./AddTopic";
