+++ src/components/teacherDashboard/subjectItem/index.ts
@@ -1,0 +1,1 @@
+export { SubjectItem } from "./SubjectItem";
