+++ src/components/teacherDashboard/subjectListCard/SubjectListCard.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from "react-i18next";
 
 import { ISubject } from "../../../utils";
 import { CardContainer } from "../cardContainer";
-import styles from "./SubjectListCard.module.css";
+import { SubjectItem } from "../subjectItem";
 
 interface ISubjectListCardProps {
     subjects: ISubject[];
@@ -47,50 +47,5 @@ export function SubjectListCard({
                 ))}
             </div>
         </CardContainer>
-    );
-}
-
-interface ISubjectItemProps {
-    subject: ISubject;
-    isSelected: boolean;
-    onSelectSubjectQuestions: (subject: ISubject) => void;
-}
-
-function SubjectItem({
-    subject,
-    isSelected,
-    onSelectSubjectQuestions,
-}: ISubjectItemProps) {
-    const { t } = useTranslation();
-
-    const getDerivedClassName = () => {
-        return isSelected
-            ? `${styles.selectedSubject} ${styles.subjectItem}`
-            : styles.subjectItem;
-    };
-
-    return (
-        <div className={getDerivedClassName()}>
-            <div>
-                <p>
-                    <span className={styles.subjectCode}>
-                        {subject.subjectCode}
-                    </span>{" "}
-                    - {subject.name}
-                </p>
-            </div>
-            {isSelected && (
-                <div>
-                    <button
-                        onClick={() => {
-                            onSelectSubjectQuestions(subject);
-                        }}
-                        className={styles.viewQuestionsBtn}
-                    >
-                        {t("teacherDashboard.viewQuestions")}
-                    </button>
-                </div>
-            )}
-        </div>
     );
 }
