+++ src/components/teacherDashboard/topicItem/index.ts
@@ -1,0 +1,1 @@
+export { TopicItem } from "./TopicItem";
