+++ src/components/teacherDashboard/topicManagerCard/TopicManagerCard.tsx
@@ -2,8 +2,8 @@ import { useTranslation } from "react-i18next";
 
 import { ISubject } from "../../../utils";
 import { CardContainer } from "../cardContainer";
-import { AddTopic } from "./AddTopic";
-import { TopicItem } from "./TopicItem";
+import { AddTopic } from "../addTopic";
+import { TopicList } from "../topicList";
 
 interface SubjectManagerCardProps {
     subject: ISubject;
@@ -22,11 +22,7 @@ export function TopicManagerCard({ subject }: SubjectManagerCardProps) {
                 {t("teacherDashboard.manageTopicsFor")} {`'${subject.name}'`}
             </h1>
             <AddTopic onCreate={handleCreate}/>
-            {subject.topics.map((topic, i) => (
-                <div key={i}>
-                    <TopicItem topic={topic} />
-                </div>
-            ))}
+            <TopicList topics={subject.topics}/>
         </CardContainer>
     );
 }
